pyaedt edb cutout bug fix (#3486)
* edb intersection bug fix

* pyaedt cutout bug fix

* Apply suggestions from code review


---------

+++ pyaedt/edb.py
@@ -1925,11 +1925,12 @@ class Edb(Database):
             elif net_name in reference_list and id not in pins_to_preserve:
                 reference_pinsts.append(i)
         for i in self.modeler.primitives:
-            net_name = i.net_name
-            if net_name not in all_list:
-                i.delete()
-            elif net_name in reference_list and not i.is_void:
-                reference_prims.append(i)
+            if i:
+                net_name = i.net_name
+                if net_name not in all_list:
+                    i.delete()
+                elif net_name in reference_list and not i.is_void:
+                    reference_prims.append(i)
         self.logger.info_timer("Net clean up")
         self.logger.reset_timer()
 
@@ -3346,29 +3347,27 @@ class Edb(Database):
         if not self.setups:
             self.siwave.add_siwave_syz_analysis()
             self.save_edb()
-        if not len(zone_primitives) == len(zone_ids):
-            self.logger.info("Number of zone primitives not equal to zone number, zone information will be lost")
-            for zone_primitive in zone_primitives:
-                edb_zone_path = os.path.join(
-                    working_directory, "{}_{}".format(zone_primitive.GetId(), os.path.basename(self.edbpath))
-                )
-                shutil.copytree(self.edbpath, edb_zone_path)
-                poly_data = zone_primitive.GetPolygonData()
-                edb_zones[edb_zone_path] = (-1, poly_data)
-        else:
-            for zone_index in range(len(zone_primitives)):
-                edb_zone_path = os.path.join(
-                    working_directory,
-                    "{}_{}".format(zone_primitives[zone_index].GetId(), os.path.basename(self.edbpath)),
-                )
-                shutil.copytree(self.edbpath, edb_zone_path)
-                poly_data = zone_primitives[zone_index].GetPolygonData()
+        for zone_primitive in zone_primitives:
+            edb_zone_path = os.path.join(
+                working_directory, "{}_{}".format(zone_primitive.GetId(), os.path.basename(self.edbpath))
+            )
+            shutil.copytree(self.edbpath, edb_zone_path)
+            poly_data = zone_primitive.GetPolygonData()
+            if self.version[0] >= 10:
+                edb_zones[edb_zone_path] = (zone_primitive.GetZoneId(), poly_data)
+            elif len(zone_primitives) == len(zone_ids):
                 edb_zones[edb_zone_path] = (zone_ids[0], poly_data)
+            else:
+                self.logger.info(
+                    "Number of zone primitives is not equal to zone number. Zone information will be lost."
+                    "Use Ansys 2024 R1 or later."
+                )
+                edb_zones[edb_zone_path] = (-1, poly_data)
         return edb_zones
 
     @pyaedt_function_handler()
-    def cutout_multizone_layout(self, zone_dict, common_reference_net):
-        """Create multizone project cutout.
+    def cutout_multizone_layout(self, zone_dict, common_reference_net=None):
+        """Create a multizone project cutout.
 
         Parameters
         ----------
@@ -3391,10 +3390,11 @@ class Edb(Database):
         """
         terminals = {}
         defined_ports = {}
+        project_connexions = None
         for edb_path, zone_info in zone_dict.items():
             edb = Edb(edbversion=self.edbversion, edbpath=edb_path)
             edb.cutout(use_pyaedt_cutout=True, custom_extent=zone_info[1], open_cutout_at_end=True)
-            if not zone_info == -1:
+            if not zone_info[0] == -1:
                 layers_to_remove = [
                     lay.name for lay in list(edb.stackup.layers.values()) if not lay._edb_layer.IsInZone(zone_info[0])
                 ]
@@ -3403,16 +3403,17 @@ class Edb(Database):
             edb.stackup.stackup_mode = "Laminate"
             edb.cutout(use_pyaedt_cutout=True, custom_extent=zone_info[1], open_cutout_at_end=True)
             edb.active_cell.SetName(os.path.splitext(os.path.basename(edb_path))[0])
-            signal_nets = list(self.nets.signal.keys())
-            defined_ports[os.path.splitext(os.path.basename(edb_path))[0]] = list(edb.excitations.keys())
-            edb_terminals_info = edb.hfss.create_vertical_circuit_port_on_clipped_traces(
-                nets=signal_nets, reference_net=common_reference_net, user_defined_extent=zone_info[1]
-            )
-            if edb_terminals_info:
-                terminals[os.path.splitext(os.path.basename(edb_path))[0]] = edb_terminals_info
+            if common_reference_net:
+                signal_nets = list(self.nets.signal.keys())
+                defined_ports[os.path.splitext(os.path.basename(edb_path))[0]] = list(edb.excitations.keys())
+                edb_terminals_info = edb.hfss.create_vertical_circuit_port_on_clipped_traces(
+                    nets=signal_nets, reference_net=common_reference_net, user_defined_extent=zone_info[1]
+                )
+                if edb_terminals_info:
+                    terminals[os.path.splitext(os.path.basename(edb_path))[0]] = edb_terminals_info
+                project_connexions = self._get_connected_ports_from_multizone_cutout(terminals)
             edb.save_edb()
             edb.close_edb()
-        project_connexions = self._get_connected_ports_from_multizone_cutout(terminals)
         return defined_ports, project_connexions
 
     @pyaedt_function_handler()
